# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "operationId": "list_courses",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "srch_tra_st_dt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "\ud6c8\ub828\uc2dc\uc791\uc77c From (YYYYMMDD \ub610\ub294 YYYY-MM-DD)",
              "title": "Srch Tra St Dt"
            },
            "description": "\ud6c8\ub828\uc2dc\uc791\uc77c From (YYYYMMDD \ub610\ub294 YYYY-MM-DD)"
          },
          {
            "name": "srch_tra_end_dt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "\ud6c8\ub828\uc2dc\uc791\uc77c To (YYYYMMDD \ub610\ub294 YYYY-MM-DD)",
              "title": "Srch Tra End Dt"
            },
            "description": "\ud6c8\ub828\uc2dc\uc791\uc77c To (YYYYMMDD \ub610\ub294 YYYY-MM-DD)"
          },
          {
            "name": "srch_tra_organ_nm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\ud6c8\ub828\uae30\uad00\uba85",
              "title": "Srch Tra Organ Nm"
            },
            "description": "\ud6c8\ub828\uae30\uad00\uba85"
          },
          {
            "name": "srch_tra_process_nm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\ud6c8\ub828\uacfc\uc815\uba85",
              "title": "Srch Tra Process Nm"
            },
            "description": "\ud6c8\ub828\uacfc\uc815\uba85"
          },
          {
            "name": "has_reg_course_man",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)",
              "default": false,
              "title": "Has Reg Course Man"
            },
            "description": "\uc218\uac15\uc2e0\uccad \uc778\uc6d0 \uc788\uc74c (regCourseMan > 0)"
          },
          {
            "name": "page_num",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page Num"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/owned-courses": {
      "get": {
        "tags": [
          "owned-courses"
        ],
        "summary": "List Owned Courses",
        "operationId": "list_owned_courses",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\uacfc\uc815\uba85 \uac80\uc0c9",
              "title": "Q"
            },
            "description": "\uacfc\uc815\uba85 \uac80\uc0c9"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "dev_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dev Year"
            }
          },
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OwnedCourseListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Create Owned Course",
        "operationId": "create_owned_course",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedCourseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedCourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/owned-courses/import/template": {
      "get": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Download Import Template",
        "operationId": "download_import_template",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/owned-courses/import": {
      "post": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Import Owned Courses",
        "operationId": "import_owned_courses",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_owned-courses-import_owned_courses"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedCourseImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/owned-courses/{course_id}": {
      "get": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Get Owned Course",
        "operationId": "get_owned_course",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedCourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Update Owned Course",
        "operationId": "update_owned_course",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnedCourseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnedCourseRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "owned-courses"
        ],
        "summary": "Delete Owned Course",
        "operationId": "delete_owned_course",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_owned-courses-import_owned_courses": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_owned-courses-import_owned_courses"
      },
...
      "CourseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page_num": {
            "type": "integer",
            "title": "Page Num"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total_count",
          "page_num",
          "page_size"
        ],
        "title": "CourseListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnedCourseCreate": {
        "properties": {
          "dev_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Year"
          },
          "dev_round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Round"
          },
          "review_round": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Round"
          },
          "division": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "ncs_dev_category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncs Dev Category"
          },
          "course_name": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Course Name"
          },
          "session_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Count"
          },
          "eval_training_volume": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Training Volume"
          },
          "result": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "grade_initial": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Initial"
          },
          "grade_23": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade 23"
          },
          "ncs_applied": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncs Applied"
          },
          "ncs_approved": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Ncs Approved"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "course_name"
        ],
        "title": "OwnedCourseCreate"
      },
      "OwnedCourseImportError": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row",
          "message"
        ],
        "title": "OwnedCourseImportError"
      },
      "OwnedCourseImportResult": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/OwnedCourseImportError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "failed",
          "errors"
        ],
        "title": "OwnedCourseImportResult"
      },
      "OwnedCourseListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "dev_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Year"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dev_year",
          "division",
          "course_name",
          "is_active"
        ],
        "title": "OwnedCourseListItem"
      },
...
      "Page_OwnedCourseListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnedCourseListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "size"
        ],
        "title": "Page[OwnedCourseListItem]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}